Separate quad key file
